pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "lxml>=6.0.2",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]
